Remote client: Move handling of Remote consoles submenu to it
That will allow to remove that menu when deleting the plugin on the fly.
