+++ spyder/plugins/remoteclient/api/__init__.py
@@ -19,5 +19,15 @@ from spyder.plugins.remoteclient.api.modules import *  # noqa
 # Max number of logged messages from the client that will be saved.
 MAX_CLIENT_MESSAGES = 1000
 
+
 class RemoteClientActions:
     ManageConnections = "manage connections"
+
+
+class RemoteClientMenus:
+    RemoteConsoles = "remote_consoles_menu"
+
+
+class RemoteConsolesMenuSections:
+    ManagerSection = "manager_section"
+    ConsolesSection = "consoles_section"
