+++ spyder/plugins/remoteclient/plugin.py
@@ -27,10 +27,13 @@ from spyder.api.plugins import OptionalPlugins, Plugins, SpyderPluginV2
 from spyder.api.translations import _
 from spyder.plugins.mainmenu.api import (
     ApplicationMenus,
+    ConsolesMenuSections,
     ToolsMenuSections,
 )
+from spyder.plugins.ipythonconsole.api import IPythonConsoleWidgetActions
 from spyder.plugins.remoteclient.api import (
     RemoteClientActions,
+    RemoteClientMenus,
 )
 from spyder.plugins.remoteclient.api.manager.base import (
     SpyderRemoteAPIManagerBase,
@@ -72,7 +75,7 @@ class RemoteClient(SpyderPluginV2):
 
     NAME = "remoteclient"
     REQUIRES = [Plugins.MainMenu]
-    OPTIONAL = [OptionalPlugins.EnvManager]
+    OPTIONAL = [OptionalPlugins.EnvManager, Plugins.IPythonConsole]
     CONF_SECTION = NAME
     CONTAINER_CLASS = RemoteClientContainer
     CONF_FILE = False
@@ -142,6 +145,8 @@ class RemoteClient(SpyderPluginV2):
         return cls.create_icon("remote_server")
 
     def on_initialize(self):
+        self._is_remote_consoles_menu_added = False
+
         self._reset_status()
         container = self.get_container()
 
@@ -161,6 +166,10 @@ class RemoteClient(SpyderPluginV2):
             container.sig_client_message_logged
         )
         self.sig_version_mismatch.connect(container.on_server_version_mismatch)
+        self.sig_connection_established.connect(
+            self._on_remote_server_connected
+        )
+        self.sig_connection_lost.connect(self._on_remote_server_disconnected)
 
     def on_first_registration(self):
         pass
@@ -192,6 +201,13 @@ class RemoteClient(SpyderPluginV2):
             before_section=ToolsMenuSections.Preferences,
         )
 
+        # Add remote console submenu
+        if (
+            self.is_plugin_available(Plugins.IPythonConsole)
+            and not self._is_remote_consoles_menu_added
+        ):
+            self._add_remote_consoles_menu()
+
     @on_plugin_teardown(plugin=Plugins.MainMenu)
     def on_mainmenu_teardown(self):
         mainmenu = self.get_plugin(Plugins.MainMenu)
@@ -200,6 +216,20 @@ class RemoteClient(SpyderPluginV2):
             RemoteClientActions.ManageConnections,
             menu_id=ApplicationMenus.Tools,
         )
+
+        mainmenu.remove_item_from_application_menu(
+            RemoteClientMenus.RemoteConsoles,
+            menu_id=ApplicationMenus.Consoles,
+        )
+
+    @on_plugin_available(plugin=Plugins.IPythonConsole)
+    def on_ipythonconsole_available(self):
+        # Add remote console submenu
+        if (
+            self.is_plugin_available(Plugins.MainMenu)
+            and not self._is_remote_consoles_menu_added
+        ):
+            self._add_remote_consoles_menu()
 
     @on_plugin_available(plugin=OptionalPlugins.EnvManager)
     def on_envmanager_available(self):
@@ -516,3 +546,32 @@ class RemoteClient(SpyderPluginV2):
 
         client = self._remote_clients[config_id]
         await client.abort_connection()
+
+    # --- For menus
+    def _add_remote_consoles_menu(self):
+        """Add remote consoles submenu to the Consoles menu."""
+        container = self.get_container()
+        container.setup_remote_consoles_submenu(render=False)
+
+        menu = container.get_menu(RemoteClientMenus.RemoteConsoles)
+        mainmenu = self.get_plugin(Plugins.MainMenu)
+        mainmenu.add_item_to_application_menu(
+            menu,
+            menu_id=ApplicationMenus.Consoles,
+            section=ConsolesMenuSections.New,
+            before=IPythonConsoleWidgetActions.ConnectToKernel,
+        )
+
+        self._is_remote_consoles_menu_added = True
+
+    def _on_remote_server_connected(self, config_id: str):
+        self.get_container().setup_server_consoles_submenu(config_id)
+
+    def _on_remote_server_disconnected(self, config_id: str):
+        # Try to reconnect any remote consoles bound to this server before
+        # altering menus.
+        ipyconsole = self.get_plugin(Plugins.IPythonConsole)
+        if ipyconsole:
+            ipyconsole.get_widget().reconnect_remote_clients(config_id)
+
+        self.get_container().clear_server_consoles_submenu(config_id)
